Add a cast for the return value of xwrite to silence -Wconversion and document the fact we are casting a count to a type of lower precision.

+++ lib/xwrite.c
@@ -50,7 +50,7 @@ xwrite(int fd, const void *buffer, size_t size)
             status = 0;
         }
     }
-    return (total < size) ? -1 : total;
+    return (total < size) ? -1 : (ssize_t) total;
 }
 
 ssize_t
